# openapi/import_v1.yaml
openapi: 3.1.0
info:
  title: Waterfall API
  version: 0.1.0
  summary: Contrat API pour l'import MS Project et la gestion des ressources
  description: |
    Contrat OpenAPI du backend Waterfall.
    Le périmètre couvre l'import MS Project, l'enrichissement des tâches,
    l'export XML et le référentiel des ressources, rôles, coûts et capacités.
servers:
  - url: http://localhost:8000
    description: Local development
...
paths:
...
  /resources/nodes:
    get:
      tags: [resources]
      summary: Lister les noeuds de ressources actifs
      operationId: listResourceNodes
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Liste des noeuds
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ResourceNodeRead'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [resources]
      summary: Creer un noeud de ressources
      operationId: createResourceNode
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResourceNodeCreate'
      responses:
        '201':
          description: Noeud cree
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceNodeRead'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/ResourceNotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /resources/nodes/{nodeId}:
    get:
      tags: [resources]
      summary: Lire un noeud de ressources
      operationId: getResourceNode
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/NodeId'
      responses:
        '200':
          description: Noeud de ressources
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceNodeRead'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/ResourceNotFound'
    patch:
      tags: [resources]
      summary: Modifier ou desactiver un noeud
      operationId: updateResourceNode
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/NodeId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResourceNodeUpdate'
      responses:
        '200':
          description: Noeud modifie
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceNodeRead'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/ResourceNotFound'
  /resources/roles:
    get:
      tags: [resources]
      summary: Lister les roles actifs
      operationId: listResourceRoles
      security:
        - bearerAuth: []
      parameters:
        - name: node_id
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
        - name: include_descendants
          in: query
          required: false
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: Liste des roles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ResourceRoleRead'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [resources]
      summary: Creer un role
      operationId: createResourceRole
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResourceRoleCreate'
      responses:
        '201':
          description: Role cree
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceRoleRead'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/ResourceNotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /resources/roles/{roleId}:
    get:
      tags: [resources]
      summary: Lire un role
      operationId: getResourceRole
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/RoleId'
      responses:
        '200':
          description: Role
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceRoleRead'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/ResourceNotFound'
    patch:
      tags: [resources]
      summary: Modifier ou desactiver un role
      operationId: updateResourceRole
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/RoleId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResourceRoleUpdate'
      responses:
        '200':
          description: Role modifie
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceRoleRead'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/ResourceNotFound'
  /resources/categories:
    get:
      tags: [resources]
      summary: Lister les categories de cout
      operationId: listCostCategories
      security:
        - bearerAuth: []
      parameters:
        - name: include_inactive
          in: query
          required: false
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: Liste des categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CostCategoryRead'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [resources]
      summary: Creer une categorie de cout
      operationId: createCostCategory
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CostCategoryCreate'
      responses:
        '201':
          description: Categorie creee
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CostCategoryRead'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
  /resources/cost-types:
    get:
      tags: [resources]
      summary: Lister les types de coût
      operationId: listCostTypes
      security:
        - bearerAuth: []
      parameters:
        - name: include_inactive
          in: query
          required: false
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: Liste des types de coût
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CostTypeRead'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [resources]
      summary: Créer un type de coût
      operationId: createCostType
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CostTypeCreate'
      responses:
        '201':
          description: Type de coût créé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CostTypeRead'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
  /resources/cost-types/{costTypeId}:
    patch:
      tags: [resources]
      summary: Modifier ou désactiver un type de coût
      operationId: updateCostType
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CostTypeId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CostTypeUpdate'
      responses:
        '200':
          description: Type de coût modifié
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CostTypeRead'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/ResourceNotFound'
  /resources/categories/{categoryId}:
    get:
      tags: [resources]
      summary: Lire une categorie de cout
      operationId: getCostCategory
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CategoryId'
      responses:
        '200':
          description: Categorie
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CostCategoryRead'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/ResourceNotFound'
    patch:
      tags: [resources]
      summary: Modifier ou desactiver une categorie
      operationId: updateCostCategory
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CategoryId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CostCategoryUpdate'
      responses:
        '200':
          description: Categorie modifiee
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CostCategoryRead'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/ResourceNotFound'
  /resources/categories/{categoryId}/rates:
    get:
      tags: [resources]
      summary: Lister les taux horaires d'une categorie
      operationId: listCostCategoryRates
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CategoryId'
      responses:
        '200':
          description: Taux horaires annuels
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CostRateRead'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/ResourceNotFound'
  /resources/rates:
    get:
      tags: [resources]
      summary: Lister tous les taux horaires
      operationId: listCostRates
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Liste des taux horaires annuels
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CostRateRead'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [resources]
      summary: Creer un taux horaire annuel
      operationId: createCostRate
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CostRateCreate'
      responses:
        '201':
          description: Taux cree
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CostRateRead'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/ResourceNotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /resources/rates/{rateId}:
    patch:
      tags: [resources]
      summary: Modifier un taux horaire annuel
      operationId: updateCostRate
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/RateId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CostRateUpdate'
      responses:
        '200':
          description: Taux modifie
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CostRateRead'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/ResourceNotFound'
  /resources/inflation:
    get:
      tags: [resources]
      summary: Lister les coefficients d'inflation
      operationId: listInflationRates
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Coefficients annuels
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/InflationRateRead'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /resources/inflation/{year}:
    put:
      tags: [resources]
      summary: Creer ou modifier un coefficient d'inflation
      operationId: putInflationRate
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Year'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InflationRateUpdate'
      responses:
        '200':
          description: Coefficient enregistre
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InflationRateRead'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /resources/capacities:
    get:
      tags: [resources]
      summary: Lister les capacites de roles
      operationId: listRoleCapacities
      security:
        - bearerAuth: []
      parameters:
        - name: role_id
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
        - name: period_start
          in: query
          required: false
          schema:
            type: string
            format: date
        - name: period_end
          in: query
          required: false
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Capacites par periode
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RoleCapacityRead'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [resources]
      summary: Creer une capacite de role
      operationId: createRoleCapacity
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoleCapacityCreate'
      responses:
        '201':
          description: Capacite creee
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleCapacityRead'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/ResourceNotFound'
  /resources/capacities/{capacityId}:
    patch:
      tags: [resources]
      summary: Modifier une capacite de role
      operationId: updateRoleCapacity
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CapacityId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoleCapacityUpdate'
      responses:
        '200':
          description: Capacite modifiee
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleCapacityRead'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/ResourceNotFound'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
...
    Unauthorized:
      description: Authentification requise
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Droits administrateur requis
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    ResourceNotFound:
      description: Ressource introuvable
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
...
    ResourceNodeCreate:
      type: object
      required: [code, name, kind]
      properties:
        code:
          type: string
          minLength: 1
          maxLength: 64
        name:
          type: string
          minLength: 1
          maxLength: 255
        parent_id:
          type: integer
          minimum: 1
          nullable: true
...
    ResourceNodeRead:
      allOf:
        - $ref: '#/components/schemas/ResourceNodeCreate'
        - type: object
          required: [id, is_active, created_at, updated_at]
          properties:
            id:
              type: integer
              minimum: 1
            is_active:
              type: boolean
            created_at:
              type: string
              format: date-time
            updated_at:
              type: string
              format: date-time
...